// packages/wizzi.site/dist/src/server/services/wizzi/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\dist\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.8
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi.site\.wizzi\server\src\services\wizzi\types.ts.ittf
    utc time: Mon, 10 May 2021 17:56:08 GMT
*/
export type WizziConfigType = { 
    repoBasePath: string;
    metaBasePath: string;
    studioBasePath: string;
    jobsBasePath: string;
    crawlBasePath: string;
    githubBasePath: string;
    cheatsheetsBasePath: string;
    siteMetaBasePath: string;
    jsonDataMetaBasePath: string;
    storeKind: string;
    plugins: string[];
    pluginsBasePath: string;
};